Refactor code
- Refactor to ensure consistency of code using CQS.
