+++ src/main/java/com/yam/app/account/presentation/AccountQueryApi.java
@@ -3,13 +3,10 @@ package com.yam.app.account.presentation;
 import com.yam.app.account.application.AccountFacade;
 import com.yam.app.account.infrastructure.AccountPrincipal;
 import com.yam.app.account.infrastructure.LoginAccount;
-import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,18 +21,6 @@ public final class AccountQueryApi {
 
     public AccountQueryApi(AccountFacade accountFacade) {
         this.accountFacade = accountFacade;
-    }
-
-    @PostMapping("/api/accounts/login")
-    public ResponseEntity<Void> login(
-        @Valid @RequestBody LoginAccountCommand request) {
-        try {
-            accountFacade.login(request);
-        } catch (IllegalStateException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-        }
-
-        return ResponseEntity.ok().build();
     }
 
     @GetMapping("/api/accounts/me")
